Page Creation and Edits to Existing
Created 'WhichEORIPage'. Removed content of 'ReportTypePage' leaving only the URL. The page did not actually press the button when continued, so it could not reach the WhichEORIPage tests. Edited the BasePage to make better use of its existing functions for this. Edited the RequestReportSpec to break it down according to how Scala Feature Specs are written.