Updating random EORI to builders
Taking the random EORI generation out from the spec and adding it to the builders, setting up a new enrolment.

This is so the randomly-generated EORI can be accessed from any test. As we will need to log in with this EORI to complete the journey from the perspective of the third party itself.